backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
PyJWT==2.8.0
bcrypt==4.1.1
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2
deepface==0.0.79
tf-keras==2.15.0
opencv-python==4.8.1.78
pillow==10.1.0
SpeechRecognition==3.10.0
pydub==0.25.1
resemblyzer==0.1.1.dev0
librosa==0.10.1
soundfile==0.12.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.0
flake8==6.1.0
email-validator==2.1.0
user-agents==2.2.0
itsdangerous==2.1.2
WTForms==3.0.1
Flask-WTF==1.1.1
APScheduler==3.10.1
redis==4.5.5
kombu==5.3.0
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
bcrypt==4.1.1

# ML/Scientific packages for advanced keystroke dynamics
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.2
pandas==2.1.1